api-gateway/requirements.txt
fastapi
uvicorn
python-jose
pydantic
httpx
pydantic-settings